+++ packages/react-native-storybook/src/getStorybook/components/index.ts
@@ -1,2 +1,1 @@
-export { default as AddonPanel } from './AddonPanel';
 export { default as ErrorBoundary } from './ErrorBoundary';
