+++ packages/react-native-storybook/src/index.ts
@@ -1,5 +1,3 @@
 export { default as getStorybook } from './getStorybook';
 export { default as openStorybook } from './openStorybook';
 export { default as registerStorybook } from './registerStorybook';
-
-// export { default as getSherloParameters } from './getSherloParameters';
